# upload/extracted/Attached-Assets-main/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Цифровой обходной лист — Digital Routing Sheet HR system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /routing-steps:
    get:
      operationId: listRoutingSteps
      tags: [routing-steps]
      summary: List routing steps (filtered by routing_sheet_id or pending for current user)
      security:
        - bearerAuth: []
      parameters:
        - name: routingSheetId
          in: query
          schema:
            type: number
        - name: pending
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of routing steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoutingStep"
  /routing-steps/{id}:
    get:
      operationId: getRoutingStep
      tags: [routing-steps]
      summary: Get a routing step by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Routing step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutingStep"
  /routing-steps/{id}/complete:
    post:
      operationId: completeRoutingStep
      tags: [routing-steps]
      summary: Mark a routing step as completed (by the assigned role)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteStepInput"
      responses:
        "200":
          description: Step completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutingStep"
        "403":
          description: Not authorized to complete this step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /routing-steps/{id}/override:
    post:
      operationId: overrideRoutingStep
      tags: [routing-steps]
      summary: Manually confirm a step (recruiter/HR override when role is absent)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OverrideStepInput"
      responses:
        "200":
          description: Step overridden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutingStep"

  # ── Candidate status page (token-based, no auth) ──────
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [routing-steps]
      summary: Request a presigned URL for photo upload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlInput"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"

  # ── Doctor Profiles ───────────────────────────────────
...
  /audit-log:
    get:
      operationId: getAuditLog
      tags: [admin]
      summary: Paginated audit log with filters (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: actorId
          in: query
          schema:
            type: number
        - name: action
          in: query
          schema:
            type: string
        - name: objectType
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Audit log page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogPage"

  # ── Admin Actions ─────────────────────────────────────
  /admin/bitrix24/test:
    post:
      operationId: testBitrix24Connection
      tags: [admin]
      summary: Test Bitrix24 connection (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bitrix24TestResult"

  # ── Termination Status Page (public) ──────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Auth ─────────────────────────────────────────────
...
    RoutingStep:
      type: object
      required: [id, routingSheetId, stepType, assignedRole, status, isBackground, createdAt]
      properties:
        id:
          type: number
        routingSheetId:
          type: number
        stepType:
          type: string
          enum: [hr_registration, marketing_photo, tb_briefing, it_accounts, audit_training, doctor_profile, site_publication, final_review]
        assignedRole:
          type: string
          enum: [admin, recruiter, hr, marketing, tb, it, audit, chief_physician, account_manager]
        assignedUserId:
          type: number
        status:
          type: string
          enum: [pending, in_progress, completed, skipped]
        isBackground:
          type: boolean
        notes:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        completedById:
          type: number
        completedByName:
          type: ["string", "null"]
        isOverride:
          type: boolean
        stepData:
          type: ["object", "null"]
          additionalProperties: true
        completedAt:
          type: ["string", "null"]
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    CompleteStepInput:
      type: object
      required: [notes]
      properties:
        notes:
          type: string
        branchId:
          type: number
        positionId:
          type: number
        photoUrl:
          type: string
        stepData:
          type: object
          additionalProperties: true
    OverrideStepInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string

    # ── Candidate status page ─────────────────────────────
...
    AuditLogEntry:
      type: object
      required: [id, actorName, action, objectType, createdAt]
      properties:
        id:
          type: number
        actorName:
          type: string
        action:
          type: string
        objectType:
          type: string
        objectId:
          type: number
        details:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ── Doctor Profiles ───────────────────────────────────
...
    AuditLogPage:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogEntry"
        total:
          type: number
        limit:
          type: number
        offset:
          type: number
    Bitrix24TestResult:
      type: object
      required: [ok, message]
      properties:
        ok:
          type: boolean
        message:
          type: string

    # ── Object storage ────────────────────────────────────
    UploadUrlInput:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
        size:
          type: number
        contentType:
          type: string
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string